backend/requirements.txt
fastapi
"uvicorn[standard]"
pydantic
python-dotenv
httpx
requests
loguru
sqlite-utils
google-generativeai
langdetect
vosk
soundfile
gTTS
pyttsx3
redis